+++ views/trangchu.ejs
@@ -26,6 +26,6 @@ document.body.onclick = function (e) {
    <h1> Website quản lý ảnh</h1>
    <div id="listImage"></div>
 </div>
-<p id ="a1">ádasdsad</p>
+<p id ="a1"></p>
 </body>
 </html>
